Struct es_runtime::esreflection::EsProxy

pub struct EsProxy { /* fields omitted */ }

Implementations

impl EsProxy

pub fn builder(
namespace: Vec<&'static str>,
class_name: &'static str
) -> EsProxyBuilder

create a builder struct to build an EsProxy

pub fn dispatch_event(
&self,
rt: &EsRuntime,
obj_id: i32,
event_name: &'static str,
event_obj: EsValueFacade
)

dispatch an event for an instance of the class

you can pass an EsValueFacade as event obj

Example

use es_runtime::esruntimebuilder::EsRuntimeBuilder;
use es_runtime::esreflection::EsProxyBuilder;
use es_runtime::esvaluefacade::EsValueFacade;

let rt = EsRuntimeBuilder::default().build();
let es_proxy = EsProxyBuilder::new(vec!["my", "biz"], "MyClass")
.constructor(|args| {
   Ok(1)
})
.event("some_event").build(&rt);
rt.eval_sync("let i = new my.biz.MyClass(); \
             i.addEventListener('some_event', (evtObj) => {\
                 console.log('it happened!');\
             });", "test_dispatch_event.es");
es_proxy.dispatch_event(&rt, 1, "some_event", EsValueFacade::undefined());

pub fn dispatch_static_event(
&self,
rt: &EsRuntime,
event_name: &'static str,
event_obj: EsValueFacade
)

dispatch an event for an instance of the class

you can pass an EsValueFacade as event obj

Example

use es_runtime::esruntimebuilder::EsRuntimeBuilder;
use es_runtime::esreflection::EsProxyBuilder;
use es_runtime::esvaluefacade::EsValueFacade;

let rt = EsRuntimeBuilder::default().build();
let es_proxy = EsProxyBuilder::new(vec!["my", "biz"], "MyClass")
.constructor(|args| {
   Ok(1)
})
.event("some_event").build(&rt);
rt.eval_sync("let i = new my.biz.MyClass(); \
             i.addEventListener('some_event', (evtObj) => {\
                 console.log('it happened!');\
             });", "test_dispatch_event.es");
es_proxy.dispatch_event(&rt, 1, "some_event", EsValueFacade::undefined());

pub fn get_canonical_name(&self) -> String

get the canonical name of the Proxy Class, this includes the namespace e.g. "my.biz.MyApp"

Example

use es_runtime::esruntimebuilder::EsRuntimeBuilder;
use es_runtime::esreflection::EsProxyBuilder;
use es_runtime::esvaluefacade::EsValueFacade;

let rt = EsRuntimeBuilder::default().build();
let es_proxy = EsProxyBuilder::new(vec!["my", "biz"], "MyClass").build(&rt);
assert_eq!(es_proxy.get_canonical_name().as_str(), "my.biz.MyClass");