Struct es_runtime::esruntime::EsRuntime

pub struct EsRuntime { /* fields omitted */ }

the EsRuntime is a facade that adds all script todo's to the EsRuntimes's event queue so they are invoked in a single worker thread you can wait for those tasks to complete by calling the _sync variants of the public methods here

Implementations

impl EsRuntime

pub fn builder() -> EsRuntimeBuilder

create a builder to instantiate an EsRuntime

pub fn start_gc_deamon(&self, interval: Duration)

start a thread which calls the cleanup method and then the garbage collector

pub fn cleanup_sync(&self)

this method should be called when you want to run the garbage collector

pub fn eval_sync(
&self,
code: &str,
file_name: &str
) -> Result<EsValueFacade, EsErrorInfo>

eval a script and wait for it to complete

pub fn load_module_sync(
&self,
module_src: &str,
module_file_name: &str
) -> Result<(), EsErrorInfo>

load a script module and run it

Example

use es_runtime::esruntimebuilder::EsRuntimeBuilder;
let rt = EsRuntimeBuilder::new().build();
rt.load_module_sync("console.log('running a module, you can import and export in and from modules');", "test_module.mes");

pub fn eval_void_sync(
&self,
code: &str,
file_name: &str
) -> Result<(), EsErrorInfo>

eval a script and wait for it to complete

pub fn call_sync(
&self,
obj_names: Vec<&'static str>,
function_name: &str,
args: Vec<EsValueFacade>
) -> Result<EsValueFacade, EsErrorInfo>

call a function by name and wait for it to complete

Example

use es_runtime::esruntimebuilder::EsRuntimeBuilder;

let rt = EsRuntimeBuilder::new().build();
rt.eval_sync("this.com = {stuff: {method: function(){console.log('my func');}}}", "test_call_sync.es").ok().expect("script failed");
rt.call_sync(vec!["com", "stuff"], "method", vec![]).ok().expect("call method failed");

pub fn eval(&self, eval_code: &str, file_name: &str)

eval a script and don't wait for it to complete

pub fn call(
&self,
obj_names: Vec<&'static str>,
function_name: &str,
args: Vec<EsValueFacade>
)

call a function by name and don't wait for it to complete

pub fn do_with_inner<R, F: FnOnce(&EsRuntimeInner) -> R>(&self, f: F) -> R

pub fn do_in_es_event_queue<J>(&self, immutable_job: J) where
J: FnOnce(&SmRuntime) + Send + 'static,

run a closure in the worker thread of this runtime's event queue, this is needed if you want to use the inner SmRuntime on which u can use the jsapi_utils

pub fn do_in_es_event_queue_sync<R: Send + 'static, J>(
&self,
immutable_job: J
) -> R where
J: FnOnce(&SmRuntime) -> R + Send + 'static,

run a closure in the worker thread of this runtime's event queue and wait for it to complete, this is needed if you want to use the inner SmRuntime on which u can use the jsapi_utils

pub fn add_helper_task<T>(task: T) where
T: FnOnce() + Send + 'static,

add a task the the "helper" thread pool

pub fn add_global_sync_function<F>(&self, name: &'static str, func: F) where
F: Fn(Vec<EsValueFacade>) -> Result<EsValueFacade, String> + Send + 'static,

add a global function to the runtime which is callable just like any other js function

Example

use es_runtime::esruntimebuilder::EsRuntimeBuilder;
use es_runtime::esvaluefacade::EsValueFacade;

let rt = EsRuntimeBuilder::new().build();
rt.add_global_sync_function("test_add_global_sync", |_args| {
     Ok(EsValueFacade::new_i32(361))
});
let esvf = rt.eval_sync("test_add_global_sync();", "test_add_global_sync_function.es").ok().expect("test_add_global_sync_function failed");
assert_eq!(esvf.get_i32(), 361);

pub fn add_global_async_function<F>(&self, name: &'static str, func: F) where
F: Fn(Vec<EsValueFacade>) -> Result<EsValueFacade, String> + Send + Sync + 'static,

add a global function to the runtime which is callable just like any other js function this async variant will run the method in a separate thread and return the result as a Promise

Example

use es_runtime::esruntimebuilder::EsRuntimeBuilder;
use es_runtime::esvaluefacade::EsValueFacade;
use std::time::Duration;

let rt = EsRuntimeBuilder::new().build();
rt.add_global_async_function("test_add_global_async", |_args| {
    Ok(EsValueFacade::new_i32(351))
});
let esvf = rt.eval_sync("test_add_global_async();", "test_add_global_async_function.es").ok().expect("test_add_global_async_function failed");
assert!(esvf.is_promise());
let prom_res = esvf.get_promise_result_blocking(Duration::from_secs(5)).ok().expect("promise timed out");
assert_eq!(prom_res.ok().expect("test_add_global_async_function failed").get_i32(), 351);