Type Definition es_runtime::esruntime::ModuleCodeLoader

type ModuleCodeLoader = dyn Fn(&str, &str) -> Option<EsScriptCode> + Send + Sync + 'static;

A ModuleCodeLoader function is used to load code into the runtime The first argument is the (relative) path of the module to import The second argument is the absolute path to the module which is importing the new module (reference_path) the EsScriptCode struct which is returned should allways contain an absolute path even if the module is loaded with a relative path