Struct es_runtime::esvaluefacade::EsValueFacade

pub struct EsValueFacade { /* fields omitted */ }

the EsValueFacade is a converter between rust variables and script objects when receiving a EsValueFacade from the script engine it's data is always a clone from the actual data so we need not worry about the value being garbage collected

Example

use es_runtime::esruntimebuilder::EsRuntimeBuilder;

let rt = EsRuntimeBuilder::default().build();
let esvf = rt.eval_sync("123", "test_es_value_facade.es").ok().unwrap();
assert!(esvf.is_i32());
assert_eq!(esvf.get_i32(), 123);

Implementations

impl EsValueFacade

pub fn undefined() -> Self

create a new EsValueFacade representing an undefined value

pub fn new_f64(num: f64) -> Self

create a new EsValueFacade representing a float

pub fn new_obj(props: HashMap<String, EsValueFacade>) -> Self

create a new EsValueFacade representing a basic object with properties as defined in the HashMap

pub fn new_i32(num: i32) -> Self

create a new EsValueFacade representing a signed integer

pub fn new_str(s: String) -> Self

create a new EsValueFacade representing a String

pub fn new_bool(b: bool) -> Self

create a new EsValueFacade representing a bool

pub fn new_array(vals: Vec<EsValueFacade>) -> Self

create a new EsValueFacade representing an Array

pub fn new_promise<C>(resolver: C) -> EsValueFacade where
C: FnOnce() -> Result<EsValueFacade, String> + Send + 'static,

create a new EsValueFacade representing a Promise, the passed closure will actually run in a seperate helper thread and resolve the Promise that is created in the script runtime

Example

use es_runtime::esruntimebuilder::EsRuntimeBuilder;
use es_runtime::esvaluefacade::EsValueFacade;
use std::time::Duration;

let rt = EsRuntimeBuilder::new().build();
rt.eval_sync("let myFunc = function(a){\
    a.then((res) => {\
        console.log('a resolved with %s', res);\
    });\
};", "test_new_promise.es");
let esvf_arg = EsValueFacade::new_promise(|| {
    // do complicated calculations or whatever here, it will run async
    // then return Ok to resolve the promise or Err to reject it
    Ok(EsValueFacade::new_i32(123))
});
rt.call_sync(vec![], "myFunc", vec![esvf_arg]);
// wait for promise to resolve
std::thread::sleep(Duration::from_secs(1));

pub fn get_string(&self) -> &str

get the String value

pub fn get_i32(&self) -> i32

get the i32 value

pub fn get_f64(&self) -> f64

get the f64 value

pub fn get_boolean(&self) -> bool

get the boolean value

pub fn is_promise(&self) -> bool

check if this esvf was a promise which was returned from the script engine

pub fn get_promise_result_blocking(
&self,
timeout: Duration
) -> Result<Result<EsValueFacade, EsValueFacade>, RecvTimeoutError>

wait for a promise to resolve in rust

Example

use es_runtime::esruntimebuilder::EsRuntimeBuilder;
use std::time::Duration;

let rt = EsRuntimeBuilder::new().build();
// run the script and fail if script fails
let esvf_prom = rt.eval_sync(
    "let p = new Promise((resolve, reject) => {setImmediate(() => {resolve(123);});}); p;",
    "test_get_promise_result_blocking.es").ok().expect("script failed");
// wait for the promise or fail on timeout
let wait_res = esvf_prom.get_promise_result_blocking(Duration::from_secs(1))
    .ok().expect("promise timed out");
// get the ok result, fail is promise was rejected
let esvf = wait_res.ok().expect("promise was rejected");
// check the result
assert_eq!(esvf.get_i32(), 123);

pub fn get_object(&self) -> &HashMap<String, EsValueFacade>

get the value as a Map of EsValueFacades, this works when the value was an object in the script engine

Example

use es_runtime::esruntimebuilder::EsRuntimeBuilder;

let rt = EsRuntimeBuilder::new().build();
let esvf = rt.eval_sync("{a: 1, b: 2};", "test_get_object.es").ok().expect("script failed");
let map = esvf.get_object();
assert!(map.contains_key("a"));
assert!(map.contains_key("b"));

pub fn get_array(&self) -> &Vec<EsValueFacade>

get the value as a Vec of EsValueFacades, this works when the value was an array in the script engine

Example

use es_runtime::esruntimebuilder::EsRuntimeBuilder;
use es_runtime::esvaluefacade::EsValueFacade;

let rt = EsRuntimeBuilder::new().build();
let esvf = rt.eval_sync("[1, 2, 3];", "test_get_array.es").ok().expect("script failed");
let arr: &Vec<EsValueFacade> = esvf.get_array();
assert_eq!(arr.len(), 3);

pub fn invoke_function(
&self,
args: Vec<EsValueFacade>
) -> Result<EsValueFacade, EsErrorInfo>

invoke the function that was returned from the script engine

Example

use es_runtime::esruntimebuilder::EsRuntimeBuilder;
use es_runtime::esvaluefacade::EsValueFacade;

let rt = EsRuntimeBuilder::new().build();
let func_esvf = rt.eval_sync("(function(a){return (a / 2);});", "test_invoke_function.es")
    .ok().expect("script failed");
// invoke the function with 18
let res_esvf = func_esvf.invoke_function(vec![EsValueFacade::new_i32(18)])
    .ok().expect("function failed");
// check that 19 / 2 = 9
let res_i32 = res_esvf.get_i32();
assert_eq!(res_i32, 9);

pub fn is_string(&self) -> bool

check if the value is a String

pub fn is_i32(&self) -> bool

check if the value is a i32

pub fn is_f64(&self) -> bool

check if the value is a f64

pub fn is_boolean(&self) -> bool

check if the value is a bool

pub fn is_object(&self) -> bool

check if the value is an object

pub fn is_array(&self) -> bool

check if the value is an array

pub fn is_function(&self) -> bool

check if the value is an function