Module es_runtime::jsapi_utils

jsapi_utils

this mod contains utilities for working with JSAPI(SpiderMonkey) unless states otherwise you can asume that all methods in this mod and submods need to be called from the event queue of the EsRuntime

Example

    use es_runtime::esruntimebuilder::EsRuntimeBuilder;
    use es_runtime::jsapi_utils;

let rt = EsRuntimeBuilder::new().build();
rt.do_in_es_event_queue_sync(|sm_rt| {
    // use jsapi_utils here
    // if you need the Runtime, JSContext or Global object (which you almost allways will)
    // you can use this method in the SmRuntime
    sm_rt.do_with_jsapi(|rt, cx, global| {
        // use jsapi_utils here
        let there_is_a_pending_exception = jsapi_utils::get_pending_exception(cx).is_some();
    })
})

Modules

arrays

big_ints

functions

handles

modules

objects

promises

reflection
Reflection

rooting

scripts

typed_arrays

Structs

EsErrorInfo
struct that represents a script exception

Functions

es_jsid_to_string

es_jsstring_to_string
convert a JSString to a rust string

es_value_to_str
convert a StringValue to a rust string

eval
eval a piece of source code in the engine

gc
call the garbage collector

get_pending_exception
see if there is a pending exception and return it as an EsErrorInfo

get_type_of
get the type of a JSVal this is the equivalent of calling typeof val in script

new_es_value_from_str
convert a str to a StringValue so it can be used in the engine

report_exception

report_exception2

set_gc_zeal_options