Function es_runtime::jsapi_utils::arrays::vec_to_array

pub fn vec_to_array<T: ToJSValConvertible>(
    context: *mut JSContext, 
    obj: MutableHandleValue, 
    vec: Vec<T>
)

convert a Vec to an Array, please note that this does not create typed arrays, use the typed_arrays mod for that

should work for all which impl the FromJSValConvertible trait like: bool u8 i8 u16 i16 u32 i32 u64 i64 f32 f64 String

Example

use es_runtime::esruntimebuilder::EsRuntimeBuilder;
use es_runtime::jsapi_utils;
use mozjs::rooted;
use mozjs::jsval::UndefinedValue;
use es_runtime::jsapi_utils::arrays::{array_to_vec, vec_to_array, get_array_length};
use mozjs::rust::HandleObject;
   
let vec = vec![2, 6, 0, 12];

let rt = EsRuntimeBuilder::new().build();
rt.do_in_es_event_queue_sync(|sm_rt| {
sm_rt.do_with_jsapi(|rt, cx, global| {
 rooted!(in (cx) let mut rval = UndefinedValue());
 vec_to_array(cx, rval.handle_mut(), vec);
 let arr_obj_handle = unsafe{HandleObject::from_marked_location(&rval.handle().get().to_object())};
 let arr_len = get_array_length(cx, arr_obj_handle).ok().expect("could not get array length");
 assert_eq!(arr_len, 4);
})
});