Module es_runtime::jsapi_utils::functions

Functions

call_function
call a function

call_function2
call a function with a rooted arguments array

call_function_name
call a function by namespace and name

call_function_name2
call a function by name with a rooted arguments array

call_function_value
call a function by value

call_function_value2
call a function by value with a rooted arguments array

call_namespace_function_name
call a function by namespace and name

call_namespace_function_name2
call a function by name on an object by name e.g. esses.cleanup() can be called by calling call_namespace_function_name(cx, glob, vec!["esses"], "cleanup", vec![]);

compile_function
compile a function

define_native_constructor
define a new native function on an object

define_native_function
define a new native function on an object

new_callback
create a new callback function based on a closure

new_callback_raw
create a new callback function based on a closure

new_native_constructor

new_native_function
define a new native function

object_is_function
check whether an Object is a function

value_is_function
check whether an Value is a function

Type Definitions

Callback
closure type which can be used to create a callback