Module es_runtime::jsapi_utils::objects

Constants

NULL_JSOBJECT

Functions

deep_freeze_object
deep freeze an Object

define_new_object

freeze_object
freeze an Object

get_constructor
get the constructor of an object

get_es_obj_prop_val
get a single member of a JSObject

get_es_obj_prop_val_as_i32
util method to quickly get a property of a JSObject as i32

get_es_obj_prop_val_as_string
util method to quickly get a property of a JSObject as String

get_es_obj_prop_val_as_string_raw
util method to quickly get a property of a JSObject as String

get_es_obj_prop_val_raw
get a single member of a JSObject

get_js_obj_prop_names
get all the propertynames of an object

get_or_define_namespace
get a namespace object and create any part that is not yet defined

get_prototype
get the prototype of an object

new_from_constructor
construct a new object based on a constructor

new_object
create a new object in the engine

new_object_from_prototype
create a new object based on a prototype object

new_object_value
create a new objectValue in the engine

set_es_obj_prop_val_permanent
set a property of an object

set_es_obj_prop_value
set a property of an object

set_es_obj_prop_value_raw
set a property of an object