Module es_runtime::jsapi_utils::reflection

Reflection

Using rust data from script

U can use the Proxy struct in jsapi_utils::reflection to create a proxy object

use mozjs::rooted;
use mozjs::rust::HandleValue;
use mozjs::jsval::{Int32Value, UndefinedValue};
use mozjs::jsapi::JSContext;
use es_runtime::jsapi_utils;
use log::debug;
use es_runtime::esruntimebuilder::EsRuntimeBuilder;

let rt = EsRuntimeBuilder::new().build();
// since we're working with jsapi we need to run in the event queue of the runtime
rt.do_in_es_event_queue_sync(|sm_rt| {
    // and then get our jsapi objects
    sm_rt.do_with_jsapi(|_rt, cx, global|{
        let proxy_arc = jsapi_utils::reflection::ProxyBuilder::new(vec![], "TestClass1")
        // if you want to create a proxy that can be constructed you need to pass a constructor
        // you need to generate an id here for your object
        // if you don't pass a constructor you can only use the static_* methods to create events, properties and methods
       .constructor(|_cx: *mut JSContext, args: Vec<HandleValue>| {
           // this will run in the sm_rt event queue so global is rooted here
           debug!("proxytest: construct");
           Ok(1)
       })
       // create a property and pass a closure for the get and set action
       .property("foo", |_cx, obj_id, mut rval| {
           debug!("proxy.get foo {}", obj_id);
           rval.set(Int32Value(123));
           Ok(())
       }, |_cx, obj_id, _val| {
           debug!("proxy.set foo {}", obj_id);
           Ok(())
       })
       // the finalizer is called when the instance is garbage collected, use this to drop your own object in rust
       .finalizer(|id: i32| {
           debug!("proxytest: finalize id {}", id);
       })
       // a method for your instance
       .method("methodA", |_cx, obj_id, args, _rval| {
           debug!("proxy.methodA called for obj {} with {} args", obj_id, args.len());
           Ok(())
       })
       // and an event that may be dispatched
       .event("saved")
       // when done build your proxy
       .build(cx, global);

       let esvf = sm_rt.eval(
           "// create a new instance of your Proxy\n\
                let tp_obj = new TestClass1('bar'); \n\
            // you can set props that are not proxied \n\
                tp_obj.abc = 1; console.log('tp_obj.abc = %s', tp_obj.abc); \n\
           // test you getter and setter\n\
               let i = tp_obj.foo; tp_obj.foo = 987; \n\
           // test your method\n\
               tp_obj.methodA(1, 2, 3); \n\
           // add an event listener\n\
                tp_obj.addEventListener('saved', (evt) => {console.log('tp_obj was saved');}); \n\
           // dispatch an event from script\n\
                tp_obj.dispatchEvent('saved', {}); \n\
           // allow you object to be GCed\n\
               tp_obj = null; i;",
           "test_proxy.es",
       ).ok().unwrap();

       assert_eq!(123, esvf.get_i32());

       // dispatch event from rust
       rooted!(in (cx) let event_obj_root = UndefinedValue());
       proxy_arc.dispatch_event(1, "saved", cx, event_obj_root.handle().into());
    });
})

Structs

Proxy
create a class def in the runtime which constructs and calls methods in a rust proxy

ProxyBuilder
the builder struct for Proxy

Functions

get_obj_id_for
retrieve the object ID for an instance of a Proxy class

get_proxy
find a ref to a proxy, use full canonical name as key, needs to run in the workerthread of the event queue

get_proxy_for
Get the Proxy of which an object is an instance

Type Definitions

Constructor

Getter

Method

Setter

StaticGetter

StaticMethod

StaticSetter