Struct es_runtime::jsapi_utils::reflection::ProxyBuilder

pub struct ProxyBuilder {
    pub namespace: Vec<&'static str>,
    pub class_name: &'static str,
    // some fields omitted
}

the builder struct for Proxy

Fields
namespace: Vec<&'static str>class_name: &'static str
Implementations

impl ProxyBuilder

Builder struct to create a Proxy object in script

pub fn new(namespace: Vec<&'static str>, class_name: &'static str) -> Self

create a new builder for a Proxy

Example

use es_runtime::esruntimebuilder::EsRuntimeBuilder;
use es_runtime::jsapi_utils::reflection::ProxyBuilder;

let rt = EsRuntimeBuilder::new().build();
rt.do_in_es_event_queue_sync(|sm_rt| {
    sm_rt.do_with_jsapi(|_rt, cx, global|{
        let _proxy = ProxyBuilder::new(vec!["com", "mybiz"], "MyClass")
        .build(cx, global);
    })
});

pub fn constructor<C>(&mut self, constructor: C) -> &mut Self where
C: Fn(*mut JSContext, Vec<HandleValue>) -> Result<i32, String> + 'static,

function to call when the Proxy is constructed please not that if you do not add a constructor you can only use the static methods, getters, setters and events

Example

use es_runtime::esruntimebuilder::EsRuntimeBuilder;
use es_runtime::jsapi_utils::reflection::ProxyBuilder;

let rt = EsRuntimeBuilder::new().build();
rt.do_in_es_event_queue_sync(|sm_rt| {
    sm_rt.do_with_jsapi(|_rt, cx, global|{
        let _proxy = ProxyBuilder::new(vec!["com", "mybiz"], "MyClass")
        .constructor(|cx, args| {
            Ok(1)
        })
        .build(cx, global);
    })
});
rt.eval_sync("let i = new com.mybiz.MyClass();",
    "test_jsapi_proxy_constructor.es")
    .ok().expect("script failed");

pub fn finalizer<F>(&mut self, finalizer: F) -> &mut Self where
F: Fn(i32) + 'static,

this closure is called when the instance of the Proxy is garbage collected, you should use this to cleanup any instances you may have created in rust

pub fn property<G, S>(
&mut self,
name: &'static str,
getter: G,
setter: S
) -> &mut Self where
G: Fn(*mut JSContext, i32, MutableHandleValue) -> Result<(), String> + 'static,
S: Fn(*mut JSContext, i32, HandleValue) -> Result<(), String> + 'static,

add a getter and setter

pub fn static_property<G, S>(
&mut self,
name: &'static str,
getter: G,
setter: S
) -> &mut Self where
G: Fn(*mut JSContext, MutableHandleValue) -> Result<(), String> + 'static,
S: Fn(*mut JSContext, HandleValue) -> Result<(), String> + 'static,

add a static getter and setter

pub fn method<M>(&mut self, name: &'static str, method: M) -> &mut Self where
M: Fn(*mut JSContext, i32, Vec<HandleValue>, MutableHandleValue) -> Result<(), String> + 'static,

add a method to the proxy class thse can be called from js as if it were members of the instance

Example

use es_runtime::esruntimebuilder::EsRuntimeBuilder;
use es_runtime::jsapi_utils::reflection::ProxyBuilder;

let rt = EsRuntimeBuilder::new().build();
rt.do_in_es_event_queue_sync(|sm_rt| {
    sm_rt.do_with_jsapi(|_rt, cx, global|{
        let _proxy = ProxyBuilder::new(vec!["com", "mybiz"], "MyClass")
        .constructor(|cx, args| {
            Ok(1)
        })
        .method("doSomething", |_cx, obj_id, _args, _rval|{
            println!("do something for obj: {}", obj_id);
            Ok(())
        })
        .build(cx, global);
    })
});
rt.eval_sync("let i = new com.mybiz.MyClass(); i.doSomething();",
    "test_jsapi_proxy_method.es")
    .ok().expect("script failed");

pub fn native_method<M>(
&mut self,
name: &'static str,
method: JSNative
) -> &mut Self

add a native method

pub fn static_method<M>(&mut self, name: &'static str, method: M) -> &mut Self where
M: Fn(*mut JSContext, Vec<HandleValue>, MutableHandleValue) -> Result<(), String> + 'static,

add a static method

pub fn static_native_method(
&mut self,
name: &'static str,
method: JSNative
) -> &mut Self

add a static native method

pub fn build(&mut self, cx: *mut JSContext, scope: HandleObject) -> Arc<Proxy>

create the proxy class, please not that this can only be used once on a builder

pub fn event(&mut self, evt_type: &'static str) -> &mut Self

define an event_type for this proxy class

pub fn static_event(&mut self, evt_type: &'static str) -> &mut Self

define a static event_type for this proxy class