Struct es_runtime::spidermonkeyruntimewrapper::SmRuntime

pub struct SmRuntime { /* fields omitted */ }

wrapper for the SpiderMonkey runtime, this struct only lives as a thread_local in the worker thread of an EsRuntime since it can only be accessed from that thread all methods here are sync

Implementations

impl SmRuntime

pub fn clone_esrt_inner(&self) -> Arc<EsRuntimeInner>

pub fn clone_current_esrt_inner_arc() -> Arc<EsRuntimeInner>

pub fn add_global_function<F>(&self, name: &'static str, func: F) where
F: Fn(*mut JSContext, CallArgs) -> bool + Send + 'static,

add a function to the global object this function will be callable from javascript just by using func_name();

Example

use es_runtime::esruntimebuilder::EsRuntimeBuilder;
use mozjs::jsval::Int32Value;
use mozjs::jsapi::CallArgs;

let rt = EsRuntimeBuilder::new().build();
rt.do_in_es_event_queue_sync(|sm_rt| {
    sm_rt.add_global_function("my_function", |_cx, args: CallArgs| {
        // impl method here
        args.rval().set(Int32Value(480));
        true
    });
});
let esvf = rt.eval_sync("my_function();", "test_add_global_function_example.es").ok().expect("test_add_global_function_example failed");
assert_eq!(esvf.get_i32(), 480);

pub fn call(
&self,
obj_names: Vec<&str>,
func_name: &str,
arguments: Vec<EsValueFacade>
) -> Result<EsValueFacade, EsErrorInfo>

pub fn load_module(
&self,
module_src: &str,
module_file_name: &str
) -> Result<(), EsErrorInfo>

load and execute a script module

pub fn eval(
&self,
eval_code: &str,
file_name: &str
) -> Result<EsValueFacade, EsErrorInfo>

eval a piece of script and return the result as a EsValueFacade

pub fn eval_void(
&self,
eval_code: &str,
file_name: &str
) -> Result<(), EsErrorInfo>

eval a piece of script and ignore the result

pub fn cleanup(&self)

run the cleanup function and run the garbage collector this also fires a pre-cleanup event in script so scripts can do a cleanup before the garbage collector runs

pub fn do_with_jsapi<C, R>(&self, consumer: C) -> R where
C: FnOnce(&Runtime, *mut JSContext, HandleObject) -> R,

use the jsapi objects in this runtime the consumer should be in the form of |rt: &Runtime, cx: *mut JSContext, global_handle: HandleObject| {} before calling the consumer the global obj is rooted and its Compartment is entered